Fix Storybook not displaying TS components and SB theme appearance
The important thing here was that I cleared the lock file when installing React. This may work with 16.4 too, but let's move on sooner rather than later. It's a backwards compatible update.

Setting appBg white makes the action panel white too, and you cannot target it for the life of it, so I updated the storybook.scss. Even though the BG is explicitly set white, the background changer still works.