Remove color from base typography
Fixes #126

I removed font-family as well. It seemed a little over-protective. I think the intention was to fight existing fonts, but if these are used on Masonite properties, what are the odds of a different font? Most websites should not have multiple font faces for their content headings anyways.

Added black text to the html/body root. Possibly overkill.