Fix BulkEdit not passing events properly
I think the original intent here was to provide the input value on the onKeyDown event since that'd get you just the value of the key pressed, but I'm sure you can get the value still. This was just a convenience fail.

Removed padding. Didn't make sense with any of the existing styles for the form controls.